% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectrodeBurstsClass.R
\name{ElectrodeBursts}
\alias{ElectrodeBursts}
\title{ElectrodeBursts Class}
\description{
This class processes MEA data related to electrode bursts, including
sample assignments and burst data. It provides methods to create raster plots
and comparison plots across different treatments.
}
\section{Super class}{
\code{\link[MEAnalysis:MEAnalysis]{MEAnalysis::MEAnalysis}} -> \code{ElectrodeBursts}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A data frame containing the processed MEA data.}

\item{\code{assignments}}{A data frame containing the sample assignments.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ElectrodeBursts-new}{\code{ElectrodeBursts$new()}}
\item \href{#method-ElectrodeBursts-find_first_occurrence}{\code{ElectrodeBursts$find_first_occurrence()}}
\item \href{#method-ElectrodeBursts-subset_by_range}{\code{ElectrodeBursts$subset_by_range()}}
\item \href{#method-ElectrodeBursts-get_sample_assignment}{\code{ElectrodeBursts$get_sample_assignment()}}
\item \href{#method-ElectrodeBursts-get_data}{\code{ElectrodeBursts$get_data()}}
\item \href{#method-ElectrodeBursts-create_raster_plot}{\code{ElectrodeBursts$create_raster_plot()}}
\item \href{#method-ElectrodeBursts-create_comparison_raster_plot}{\code{ElectrodeBursts$create_comparison_raster_plot()}}
\item \href{#method-ElectrodeBursts-clone}{\code{ElectrodeBursts$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="MEAnalysis" data-topic="MEAnalysis" data-id="create_combined_t_test_plot"><a href='../../MEAnalysis/html/MEAnalysis.html#method-MEAnalysis-create_combined_t_test_plot'><code>MEAnalysis::MEAnalysis$create_combined_t_test_plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MEAnalysis" data-topic="MEAnalysis" data-id="get_electrode_averages"><a href='../../MEAnalysis/html/MEAnalysis.html#method-MEAnalysis-get_electrode_averages'><code>MEAnalysis::MEAnalysis$get_electrode_averages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MEAnalysis" data-topic="MEAnalysis" data-id="get_sample_assignments"><a href='../../MEAnalysis/html/MEAnalysis.html#method-MEAnalysis-get_sample_assignments'><code>MEAnalysis::MEAnalysis$get_sample_assignments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MEAnalysis" data-topic="MEAnalysis" data-id="get_treatment_averages"><a href='../../MEAnalysis/html/MEAnalysis.html#method-MEAnalysis-get_treatment_averages'><code>MEAnalysis::MEAnalysis$get_treatment_averages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MEAnalysis" data-topic="MEAnalysis" data-id="get_treatment_list"><a href='../../MEAnalysis/html/MEAnalysis.html#method-MEAnalysis-get_treatment_list'><code>MEAnalysis::MEAnalysis$get_treatment_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MEAnalysis" data-topic="MEAnalysis" data-id="get_well_averages"><a href='../../MEAnalysis/html/MEAnalysis.html#method-MEAnalysis-get_well_averages'><code>MEAnalysis::MEAnalysis$get_well_averages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MEAnalysis" data-topic="MEAnalysis" data-id="perform_t_tests"><a href='../../MEAnalysis/html/MEAnalysis.html#method-MEAnalysis-perform_t_tests'><code>MEAnalysis::MEAnalysis$perform_t_tests()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MEAnalysis" data-topic="MEAnalysis" data-id="treatment_averages_t_test_plot"><a href='../../MEAnalysis/html/MEAnalysis.html#method-MEAnalysis-treatment_averages_t_test_plot'><code>MEAnalysis::MEAnalysis$treatment_averages_t_test_plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectrodeBursts-new"></a>}}
\if{latex}{\out{\hypertarget{method-ElectrodeBursts-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ElectrodeBursts Object

This method initializes the ElectrodeBursts class with a specified file path.
It reads and processes the MEA data and assignments from the file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectrodeBursts$new(filepath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filepath}}{The path to the file to be processed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class ElectrodeBursts.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectrodeBursts-find_first_occurrence"></a>}}
\if{latex}{\out{\hypertarget{method-ElectrodeBursts-find_first_occurrence}{}}}
\subsection{Method \code{find_first_occurrence()}}{
Find First Occurrence

This method finds the first occurrence of a specific substring in the data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectrodeBursts$find_first_occurrence(df, substring_to_find)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{The data frame to search in.}

\item{\code{substring_to_find}}{The substring to search for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The row number of the first occurrence or NA if not found.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectrodeBursts-subset_by_range"></a>}}
\if{latex}{\out{\hypertarget{method-ElectrodeBursts-subset_by_range}{}}}
\subsection{Method \code{subset_by_range()}}{
Subset by Range

This method subsets a data frame based on the specified row range.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectrodeBursts$subset_by_range(df, start_row, end_row)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{The data frame to subset.}

\item{\code{start_row}}{The starting row number.}

\item{\code{end_row}}{The ending row number.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A subset of the data frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectrodeBursts-get_sample_assignment"></a>}}
\if{latex}{\out{\hypertarget{method-ElectrodeBursts-get_sample_assignment}{}}}
\subsection{Method \code{get_sample_assignment()}}{
Get Sample Assignments

This method returns the processed sample assignments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectrodeBursts$get_sample_assignment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame with sample assignments.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectrodeBursts-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-ElectrodeBursts-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get MEA Data

This method returns the processed MEA data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectrodeBursts$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame with MEA data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectrodeBursts-create_raster_plot"></a>}}
\if{latex}{\out{\hypertarget{method-ElectrodeBursts-create_raster_plot}{}}}
\subsection{Method \code{create_raster_plot()}}{
Create Raster Plot

This method creates a raster plot for a specified treatment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectrodeBursts$create_raster_plot(treatment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{treatment}}{The treatment name to be plotted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object representing the raster plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectrodeBursts-create_comparison_raster_plot"></a>}}
\if{latex}{\out{\hypertarget{method-ElectrodeBursts-create_comparison_raster_plot}{}}}
\subsection{Method \code{create_comparison_raster_plot()}}{
Create Comparison Raster Plot

This method creates a comparison raster plot for multiple treatments
including a control group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectrodeBursts$create_comparison_raster_plot(
  control_group,
  treatments_array,
  plot_title
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{control_group}}{The control group to be included in the comparison.}

\item{\code{treatments_array}}{An array of treatment names to compare.}

\item{\code{plot_title}}{The title of the plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggpubr object representing the combined raster plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectrodeBursts-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ElectrodeBursts-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectrodeBursts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
